+++ AnimatedTokenMaker/Services/IEncoderService.cs
@@ -1,0 +1,9 @@
+﻿using AnimatedTokenMaker.Source;
+
+namespace AnimatedTokenMaker.Services
+{
+    public interface IEncoderService
+    {
+        void EncodeFolderAsWebm(string outputFile, string sourceFolder, ISourceSetting _sourceSetting);
+    }
+}