+++ AnimatedTokenMaker/Services/IFFmpegService.cs
@@ -1,0 +1,8 @@
+﻿using AnimatedTokenMaker.Services;
+
+namespace AnimatedTokenMaker.Services
+{
+    public interface IFFmpegService : IDecoderService, IEncoderService, ITokenMakerService
+    {
+    }
+}