+++ AnimatedTokenMaker/SourceFactory.cs
@@ -1,40 +1,55 @@
-﻿using AnimatedTokenMaker.Source;
-using System.IO;
-using System.Linq;
+﻿using AnimatedTokenMaker.Services;
+using AnimatedTokenMaker.Source;
+using System;
+using System.Collections.Generic;
 
 namespace AnimatedTokenMaker
 {
     public class SourceFactory
     {
-        private IFFmpegService _ffmpegService;
         private ISourceSetting _defaultSetting;
+        private delegate ISourceFile NewSourceFactory(string file);
 
-        public SourceFactory(IFFmpegService ffmpegService, ISourceSetting defaultSetting)
+        private Dictionary<string[], NewSourceFactory> _decoderServices;
+        private NewSourceFactory _default;
+
+        public SourceFactory(ISourceSetting defaultSetting)
         {
-            _ffmpegService = ffmpegService;
+            _decoderServices = new Dictionary<string[], NewSourceFactory>
+            {
+                { new []{ "png", "bmp", "jpg", "tiff", "wmf", "exif", "emf", "ico" }, (file) => new StaticImageSource(file) },
+                { new []{ "webp" }, (file) => CreateWebmuxSource(file) }
+            };
+
+            _default = (file) => new VideoSource(file, ServiceManager.Instance.FFmpegService, _defaultSetting);
             _defaultSetting = defaultSetting;
+        }
+
+        private ISourceFile CreateWebmuxSource(string file)
+        {
+            if (ServiceManager.Instance.WebmuxService.IsReady())
+            {
+                return new VideoSource(file, ServiceManager.Instance.WebmuxService, _defaultSetting);
+            }
+            else
+            {
+                throw new SourceNotReadyException(ServiceManager.Instance.WebmuxService.Message);
+            }
         }
 
         public ISourceFile GetSource(string file)
         {
-            ISourceFile layer;
-
-            if (IsStaticImage(file))
+            foreach (var decoder in _decoderServices)
             {
-                layer = new StaticImageSource(file);
-            }
-            else
-            {
-                layer = new VideoSource(file, _ffmpegService, _defaultSetting);
+                if (FileHelpers.HasExtension(file, decoder.Key))
+                {
+                    return decoder.Value.Invoke(file);
+                }
             }
 
-            return layer;
+            return _default.Invoke(file);
         }
 
-        private bool IsStaticImage(string file)
-        {
-            var staticExtensions = new[] { "png", "bmp", "jpg", "tiff", "wmf", "exif", "emf", "ico" };
-            return staticExtensions.Contains(Path.GetExtension(file).Trim('.'));
-        }
+
     }
 }